generate sound effects from image/requirements.txt
gradio==4.0.0
torch==2.1.2
diffusers
transformers
scipy
spaces
python-dotenv